calyx/pyproject.toml
[project]
name = "calyx"
version = "0.1.0"
description = ""
authors = [
    {name = "almagribis",email = "almagribisadli@gmail.com"}
]
readme = "README.md"
packages = [{include = "calyx"}]
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (>=1.0.5)",
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-huggingface (==1.0.1)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools.packages.find]
where = ["."]
include = ["calyx"]
namespaces = false